Emit the approximate future L2 block as the vote end on L2
We calculate the approximate L2 block that will be produced at the time
when the L1 arrives at the proposal end block. We emit this on L2 when
the proposal is bridged for the sake of integration with third party tools,
which use the data emitted by these events to determine when to close the
voting period. This is somewhat of a hack, and could/should be removed if
we get proper support from clients for L2 Flex Voting. Note that this
implementation simply hardcodes the current average block times for our
target MVP networks, namely Mainnet + Optimism.